requirements.txt
gpxpy==1.1.1
piexif==1.0.5
python-dateutil
pytz
